+++ src/extensions/yfm/YfmTable/YfmTableSpecs/index.ts
@@ -1,4 +1,4 @@
-import log from '@diplodoc/transform/lib/log.js';
+import {log} from '@diplodoc/transform/lib/log.js';
 import yfmTable from '@diplodoc/transform/lib/plugins/table/index.js';
 
 import type {ExtensionWithOptions} from '../../../../core';